refactor: Proper code structure
