+++ src/main/java/org/uom/lefterisxris/codetour/other/AskQuestionAction.java
@@ -1,4 +1,4 @@
-package org.uom.lefterisxris.codetour;
+package org.uom.lefterisxris.codetour.other;
 
 import com.intellij.ide.BrowserUtil;
 import com.intellij.openapi.actionSystem.AnAction;
