+++ src/main/java/org/uom/lefterisxris/codetour/other/PopupDialogAction.java
@@ -1,4 +1,4 @@
-package org.uom.lefterisxris.codetour;
+package org.uom.lefterisxris.codetour.other;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
