+++ src/main/java/org/uom/lefterisxris/codetour/tours/TourScanner.java
@@ -4,6 +4,7 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.startup.StartupActivity;
 import org.jetbrains.annotations.NotNull;
 import org.uom.lefterisxris.codetour.tours.state.StateManager;
+import org.uom.lefterisxris.codetour.tours.ui.ToursNotification;
 
 public class TourScanner implements StartupActivity.Background {
