+++ src/main/java/org/uom/lefterisxris/codetour/tours/ui/ToursNotification.java
@@ -1,4 +1,4 @@
-package org.uom.lefterisxris.codetour.tours;
+package org.uom.lefterisxris.codetour.tours.ui;
 
 
 import com.intellij.notification.Notification;
